requirement.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
tqdm>=4.65.0
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.41.0,<5
timm>=0.9.0
git+https://github.com/facebookresearch/sam2.git
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.22.0
ultralytics>=8.3.198
